Make finish-login error status code to 403 Forbidden

+++ tests/docker-entry.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-set -e
 
 echo '127.0.0.1 sp.example.com idp.example.com' >> /etc/hosts
 
@@ -10,11 +9,14 @@ rsync -avz /usr/local/ngx-lua-saml-sp-test/shdict_store/etc/nginx/ /etc/nginx/
 nginx -g 'daemon off;' &
 nginx_pid=$!
 
-while ! timeout 1 bash -c "echo > /dev/tcp/localhost/443" 2> /dev/null; do   
+while ! timeout 1 bash -c "echo > /dev/tcp/localhost/443" 2> /dev/null; do
   sleep 1
 done
 export LUA_PATH='/usr/local/luajit-http-client/lib/?.lua;/usr/local/luajit-http-client/vendor/?.lua;/usr/local/lbase64/?.lua;;'
 luajit test.lua
+if [ $? -ne 0 ]; then
+  echo 'shdict_store test failed!!!'
+fi
 
 kill $nginx_pid
 
@@ -27,11 +29,20 @@ rsync -avz /usr/local/ngx-lua-saml-sp-test/jwt_store/etc/nginx/ /etc/nginx/
 nginx -g 'daemon off;' &
 nginx_pid=$!
 
-while ! timeout 1 bash -c "echo > /dev/tcp/localhost/443" 2> /dev/null; do   
+while ! timeout 1 bash -c "echo > /dev/tcp/localhost/443" 2> /dev/null; do
   sleep 1
 done
 export LUA_PATH='/usr/local/luajit-http-client/lib/?.lua;/usr/local/luajit-http-client/vendor/?.lua;'
-luajit test.lua
+luajit test.lua "$@"
+if [ $? -eq 0 ]; then
+    kill $nginx_pid
+else
+    tail -n 20 /var/log/nginx/error.log
+    cat <<EOF
 
-kill $nginx_pid
-#wait $nginx_pid
+jwt_store test failed!!!
+You can check log files by running docker exec -it \$container_id bash on another terminal,
+or press Ctrl-C to stop this container.
+EOF
+    wait $nginx_pid
+fi
